+++ H6Game/H6Game.Test/H6Game.TestDistributedNetCore/TestBenckmark.cs
@@ -46,7 +46,7 @@ namespace H6Game.TestDistributedNetCore
 
         public static async Task Call(Network network)
         {
-            await network.CallMessageAsync<TestMessage, TestMessage>(send, (int)TestCMD.TestCmd);
+            await network.CallMessageAsync<TestMessage, TestMessage>(send, TestCMD.TestCmd);
 
             Count++;
             if (stopWatch.ElapsedMilliseconds > 1000)
